+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 
 #### Features supported with current release:
 - Deployment plan
+- Storage pool
 
 # 4.5.0
 #### Notes