+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # 4.6.0 (Unreleased)
 #### Notes
 Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
-- Plan Scripts
 
 #### Features supported with current release:
 - Deployment plan
+- Plan Scripts
 
 # 4.5.0
 #### Notes