+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 - Deployment plan
 - Interconnect
 - Network set
+- OS Deployment Plans
 - Storage pool
 - Storage volume