+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 - Deployment plan
 - Interconnect
 - Network set
-- Plan Scripts
+- Plan script
 - Storage pool
 - Storage volume