+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 
 #### Features supported with current release:
 - Deployment plan
+- Golden image
 - Interconnect
 - Network set
 - Storage pool