+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 - Golden image
 - Interconnect
 - Network set
+- OS deployment plan
 - Storage pool
 - Storage volume